backend/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.2
faster-whisper==1.1.0
requests>=2.28.0
numpy==1.26.4
coloredlogs==15.0.1
